test: connect to PostgreSQL only corporation table
