+++ job-a-dream-backend/domain/jobs/jobs_router.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+from connector import get_db
+from domain.corporations import corp_crud, corp_schemas
+from typing import List
+
+router = APIRouter()
